Avoid drift of defaultDate due to locale
There is no concept of a pure date in js, only absolute timestamps.
A call to `new Date(value)` with a `value` of a string will attempt
to parse a datetime from that string into an absolute and localised
timestamp. Depending on the client locale this can result in the
defaultDate being misrepresented. To counter for this we add the
locale timezone offset.